Bigmac/teleport slo implementation (#162)
* Add bigmac team's SLO configurations for Teleport service

* Add process_state SLO configuration for Teleport service

* Delete areas/empowerment/teams/bigmac/teleport/.DS_Store

* Delete areas/empowerment/teams/.DS_Store

* Create teleport_cache_stale_events_slo.yaml

* Create audit_percentage_disk_space_used_slo.yaml

The errorQuery calculates the average disk usage over a 30-day period and triggers an alert if the average is greater than 80%.

We can adjust this metrics based on our needs.

* Create auth_generate_latency_slo.yaml

The errorQuery calculates the 99th percentile latency for authentication requests. If the latency exceeds 0.2 seconds (200 milliseconds), it's considered an error. 
(We may  need to adjust the 0.2 threshold based on the actual performance targets for our Teleport authentication requests.)

The totalQuery counts all authentication request events over a 5-minute period for context.

* Create backend_read_seconds_bucket.yaml

The errorQuery uses the histogram_quantile function to calculate the 95th percentile of latencies from the backend_read_seconds_bucket metric. It flags an error if the latency is greater than 50 milliseconds (0.05 seconds). 

We should adjust the threshold to suit your performance criteria.

* Create backend_write_latency_slo.yaml

The errorQuery calculates the 95th percentile latency and triggers an alert if it exceeds 100ms (0.1 seconds).

The totalQuery counts the total backend write operations over the last 5 minutes.

budgetingMethod: "occurrences" tracks the number of times the SLO is breached.

* Create dynamodb_latency_slo.yaml

The errorQuery checks if the 99th percentile latency over the last 5 minutes is greater than 200 milliseconds.

The totalQuery counts the total number of DynamoDB request operations over the last 5 minutes.

The budgetingMethod: "occurrences" keeps track of how often this SLO is not met.

* Delete areas/.DS_Store

* Delete areas/empowerment/.DS_Store

* Delete areas/empowerment/teams/bigmac/.DS_Store

* Delete areas/empowerment/teams/bigmac/dns directory

* Delete areas/empowerment/teams/bigmac/cni directory

* Update audit_percentage_disk_space_used_slo.yaml

* Delete .DS_Store

* Refactored the yaml file. Also ensured that alerts are only sent during working hours. This improves the alerting system by providing immediate access to resolution steps and reducing noise from alerts outside of working hours.

* Update README.md

* bigmac team folder added in areas/kaas/teams.

* Delete areas/kaas/bigmac/teleport directory

* Update SLO queries in YAML files

* Update SLO configurations for Teleport services

* Update audit_percentage_disk_space_used_slo.yaml

* Update auth_generate_latency_slo.yaml

* Update backend_write_latency_slo.yaml

* Update backend_read_seconds_bucket.yaml

* Update auth_generate_latency_slo.yaml

* helm-template generate

* Update SLO configurations for Teleport services

* Delete .DS_Store

* Delete areas/.DS_Store

* Delete areas/kaas/.DS_Store

* Update SLO configurations for Teleport services

* Update SLO configurations for Teleport services

* Update audit_percentage_disk_space_used_slo.yaml

* Update auth_generate_latency_slo.yaml

* Update auth_generate_latency_slo.yaml

* Update backend_read_seconds_bucket.yaml

* Update backend_write_latency_slo.yaml

* Update dynamodb_latency_slo.yaml

* Update process_state_slo.yaml

* Update .gitignore

* Update areas/kaas/teams/bigmac/teleport/slos/teleport_cache_stale_events_slo.yaml

* Update areas/kaas/teams/bigmac/teleport/slos/teleport_cache_stale_events_slo.yaml

---------

Co-authored-by: Onur YILMAZ <onuryilmaz@Onurs-MacBook-Pro.local>
Co-authored-by: Antonia <56017655+anvddriesch@users.noreply.github.com>
Co-authored-by: Onur YILMAZ <onuryilmaz@r143090010.stusta.swh.mhn.de>
Co-authored-by: Quentin Bisson <quentin@giantswarm.io>